dominium-56 Command Gateway und Query Gateway zur Wiederverwendung vorbereiten Refactoring: Command Gateway benutzt Redux Pattern
