+++ ui/frontend/src/main/src/app/app.component.spec.ts
@@ -6,7 +6,6 @@ import { AppModule } from './app.module';
 import { APP_BASE_HREF } from '@angular/common';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { QueryGatewayService } from './shared/query-gateway/query-gateway.service';
-import { RootEpicsService } from './store/root-epics.service';
 import { CommandGatewayModule } from './shared/command-gateway/command-gateway.module';
 
 describe('AppComponent', () => {
