+++ ui/frontend/src/main/src/app/shared/command-gateway/command-gateway.module.ts
@@ -1,22 +1,21 @@
+import { HttpClient } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { CommandGatewayActionsService } from './command-gateway-actions.service';
-import { CommandGatewayEpicsService } from './command-gateway-epics.service';
+import { EpicProvider } from '../redux-utils/provider';
 import { CommandGatewayService } from './command-gateway.service';
+import { fallsCommandAngefordert } from './command.redux';
 import { IdGeneratorService } from './id-generator.service';
 import { CommandProgressComponent } from './command-progress/command-progress.component';
 import { AppMaterialModule } from '../app-material-module';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { CovalentDialogsModule } from '@covalent/core';
+import { CovalentDialogsModule, TdDialogService } from '@covalent/core';
 
 @NgModule({
     providers: [
-        CommandGatewayActionsService,
-        CommandGatewayEpicsService,
         CommandGatewayService,
         IdGeneratorService,
-
+        EpicProvider(fallsCommandAngefordert, [HttpClient, TdDialogService])
     ],
     imports: [
         CommonModule,
