+++ ui/frontend/src/main/src/app/shared/command-gateway/command-gateway.service.spec.ts
@@ -4,7 +4,6 @@ import { CommandGatewayService } from './command-gateway.service';
 import { NgReduxModule } from '@angular-redux/store';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { StoreModule } from '../../store/store.module';
-import { RootEpicsService } from '../../store/root-epics.service';
 import { CommandGatewayModule } from './command-gateway.module';
 import { CommandType } from '../../inventur/command-type';
 import { AppComponent } from '../../app.component';
@@ -27,8 +26,9 @@ describe('CommandGatewayService', () => {
                 StoreModule,
                 CommandGatewayModule
             ],
-            providers: [RootEpicsService,
-                {provide: APP_BASE_HREF, useValue: '/'}]
+            providers: [
+                {provide: APP_BASE_HREF, useValue: '/'}
+            ]
         });
     });
