+++ ui/frontend/src/main/src/app/store/inventur-reducers.ts
@@ -1,6 +1,6 @@
+import { CommandGatewayActionType } from '../shared/command-gateway/command.redux';
 import { QueryGatewayActionType } from '../shared/query-gateway/query.redux';
 import { INVENTUR_INITIAL_STATE, InventurState } from './model';
-import { CommandGatewayActionType } from '../shared/command-gateway/command-gateway-actions.service';
 import { CommandType } from '../inventur/command-type';
 import { QueryType } from '../inventur/query-type';
