+++ ui/frontend/src/main/src/app/store/reducers.ts
@@ -1,4 +1,4 @@
-import { commandReducer } from '../shared/command-gateway/command-reducers';
+import { command } from '../shared/command-gateway/command.redux';
 import { AppState } from './model';
 import { combineReducers } from 'redux';
 import { inventurReducer } from './inventur-reducers';
@@ -7,6 +7,6 @@ import { query } from '../shared/query-gateway/query.redux';
 
 export const rootReducer = combineReducers<AppState>({
     inventur: inventurReducer,
-    command: commandReducer,
+    command: command,
     query: query
 });
