--- ui/frontend/src/main/src/app/store/root-epics.service.ts
@@ -1,13 +1,0 @@
-import { Injectable } from '@angular/core';
-import { CommandGatewayEpicsService } from '../shared/command-gateway/command-gateway-epics.service';
-
-@Injectable()
-export class RootEpicsService {
-
-    constructor(private commandBusEpics: CommandGatewayEpicsService) {
-    }
-
-    public createEpics() {
-        return [...this.commandBusEpics.createEpic()];
-    }
-}
