+++ ui/frontend/src/main/src/app/store/store.module.ts
@@ -3,22 +3,20 @@ import { CommonModule } from '@angular/common';
 import { DevToolsExtension, NgRedux, NgReduxModule } from '@angular-redux/store';
 import { Action } from 'redux';
 import { createEpicMiddleware, Epic, EpicMiddleware } from 'redux-observable';
+import { CommandGatewayModule } from '../shared/command-gateway/command-gateway.module';
 import { REDUX_EPIC } from '../shared/redux-utils/provider';
 import { QueryGatewayModule } from '../shared/query-gateway/query-gateway.module';
 import { QueryGatewayService } from '../shared/query-gateway/query-gateway.service';
 import { APP_INITIAL_STATE, AppState } from './model';
 import { rootReducer } from './reducers';
-import { RootEpicsService } from './root-epics.service';
 
 @NgModule({
-    imports: [CommonModule, NgReduxModule, QueryGatewayModule],
+    imports: [CommonModule, NgReduxModule, QueryGatewayModule, CommandGatewayModule],
     exports: [NgReduxModule],
-    providers: [RootEpicsService],
     declarations: []
 })
 export class StoreModule {
     constructor(store: NgRedux<AppState>,
-                rootEpics: RootEpicsService,
                 @Inject(REDUX_EPIC) @Optional() epics: Epic<Action, {}>[] = [],
                 devTools: DevToolsExtension) {
 
@@ -28,7 +26,7 @@ export class StoreModule {
         store.configureStore(
             rootReducer,
             APP_INITIAL_STATE,
-            [...rootEpics.createEpics(), ...middleware],
+            [...middleware],
             devTools.isEnabled() ? [devTools.enhancer()] : [])
     }
 }
